Fixed debug UI incorrect bounds rectangle.
