+++ MyPiUI/Drawing/Buffers/IDrawBuffer.cs
@@ -10,7 +10,6 @@ public interface IDrawBuffer
     public void Clear(Rectangle region);
     public void Clear(Color color, Rectangle region);
 
-    public void ClearDirtyRegions();
     public void SetClipRect(Rectangle region);
     public void ClearClipRect();
 
@@ -19,7 +18,7 @@ public interface IDrawBuffer
     public void DrawRect(Rectangle rect, Color color);
     public void FillRect(Rectangle rect, Color color);
 
-    public void DrawImage(Point point, Span<byte> image);
+    public void DrawImage(Rectangle rect, Span<byte> image);
 
     public void DrawText(Point position, string text, string fontFamily, float fontSize, Color color);
     public Size MeasureText(string text, string fontFamily, float fontSize);
