+++ MyPiUI/UI/Controls/TextUIElement.cs
@@ -1,8 +1,5 @@
 using System.Xml.Serialization;
 
-using MyPiUI.Drawing;
-using MyPiUI.Drawing.Buffers;
-
 namespace MyPiUI.UI.Controls;
 
 public abstract class TextUIElement : UIElement
@@ -37,30 +34,4 @@ public abstract class TextUIElement : UIElement
 
     [XmlIgnore]
     public int FontSize { get; set; }
-
-    private IDrawBuffer? _buffer;
-
-    public override void Init(MyGraphicsContext graphicsContext, IDrawBuffer buffer)
-    {
-        _buffer = buffer;
-        
-        CalculateBounds();
-        
-        base.Init(graphicsContext, buffer);
-    }
-
-    public void CalculateBounds()
-    {
-        if (string.IsNullOrWhiteSpace(Text) ||
-            string.IsNullOrWhiteSpace(FontFamily) ||
-            _buffer is null)
-        {
-            return;
-        }
-
-        var size = _buffer.MeasureText(Text, FontFamily, FontSize);
-        
-        Width = size.Width;
-        Height = size.Height;
-    }
 }
